Refactor the Java acceptance test.
This patch pulls out Java code in separate standalone .java files.
