+++ acceptance/java/main.java
@@ -1,0 +1,8 @@
+import java.sql.*;
+
+public class main {
+    public static void main(String[] args) throws Exception {
+        Connection conn = TestConnection.getConnection();
+        TestStatements.runTests(conn)
+    }
+}
